Flask/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
sqlparse
